refactor

+++ gophoenix/core/store/keystore.go
@@ -39,7 +39,7 @@ func (self *KeyStore) Unlock(phrase string) error {
 func (self *KeyStore) SignTx(tx *types.Transaction, chainID int64) (*types.Transaction, error) {
 	return self.KeyStore.SignTx(
 		self.GetAccount(),
-		tx, big.NewInt(chainID),
+		tx, big.NewInt(int64(chainID)),
 	)
 }
